requirements.txt
scikit-learn>=1.0.2
django>=4.2.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
joblib>=1.1.0
pandas>=1.5.0
gunicorn>=20.1.0
whitenoise>=6.0.0
dj-database-url>=0.5.0
psycopg2-binary>=2.9.3 